requirements.txt
# Python Dependencies
pyspark==3.5.0
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Iceberg
pyiceberg==0.6.1
pyarrow==14.0.2

# Database & Storage
minio==7.2.0
clickhouse-driver==0.2.6
pymilvus==2.3.4
opensearch-py==2.4.0

# API & Web
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.2

# Social Media APIs
praw==7.7.1
tweepy==4.14.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.4
pytest-spark==0.6.0

# Development
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.1
python-dotenv==1.0.0